backend/requirements.api.txt
# ===============================================
# requirements.api.txt - FastAPI Backend
# ===============================================

# FastAPI Core
uvicorn[standard]==0.24.0
fastapi==0.104.1

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Pydantic and settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Machine Learning 
numpy==1.25.2
scikit-learn==1.3.2

# OpenAI
openai==1.8.0
anthropic>=0.7.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# HTTP and networking
httpx==0.26.0
requests==2.31.0

# Data processing
pandas==2.1.4

# Compatible with Alibaba-NLP/gte-multilingual-base requirements
huggingface_hub==0.19.4
prometheus_client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

boto3==1.34.0
botocore==1.34.0

assemblyai>=0.20.0